move files into factory pattern
