+++ .env.development
@@ -1,0 +1,4 @@
+env_filename=.env.development
+
+DATABASE_NAME=portfolio_manager_dev
+SQLALCHEMY_TRACK_MODIFICATIONS=1
