+++ .env.production
@@ -1,0 +1,4 @@
+env_filename=.env.production
+
+DATABASE_NAME=portfolio_manager_production
+SQLALCHEMY_TRACK_MODIFICATIONS=0
