+++ .env.test
@@ -1,0 +1,4 @@
+env_filename=.env.test
+
+DATABASE_NAME=portfolio_manager_test
+SQLALCHEMY_TRACK_MODIFICATIONS=0
