+++ api/schema.graphql
@@ -17,30 +17,30 @@ type GroupTrade {
 type Trade {
 	id: ID!,
 	symbol: String!,
-	tradeID: String!,
-	ibOrderID: String!,
+	tradeID: String,
+	ibOrderID: String,
 	ibExecID: String,
-	transactionID: String!,
-	accountId: String!,
-	underlyingSymbol: String!,
-	description: String!,
-	multiplier: Int!,
-	strike: String!,
-	expiry: String!,
-	putCall: String!,
-	dateTime: String!,
-	tradeDate: String!,
-	quantity: Float!,
-	tradePrice: Float!,
-	proceeds: Float!,
-	ibCommission: Float!,
-	netCash: Float!,
-	closePrice: Float!,
-	openCloseIndicator: String!,
-	fifoPnlRealized: Float!,
-	fxPnl: Float!,
-	mtmPnl: Float!,
-	buySell: String!
+	transactionID: String,
+	accountId: String,
+	underlyingSymbol: String,
+	description: String,
+	multiplier: Int,
+	strike: String,
+	expiry: String,
+	putCall: String,
+	dateTime: String,
+	tradeDate: String,
+	quantity: Float,
+	tradePrice: Float,
+	proceeds: Float,
+	ibCommission: Float,
+	netCash: Float,
+	closePrice: Float,
+	openCloseIndicator: String,
+	fifoPnlRealized: Float,
+	fxPnl: Float,
+	mtmPnl: Float,
+	buySell: String
 	
 	# -- accountId
 	# acctAlias
