+++ api/static/js/TradesPage.js
