+++ test/test_app.py
@@ -1,0 +1,11 @@
+from test.app_test_case import AppTestCase
+
+
+class AppTest(AppTestCase):
+    def test_root_path(self):
+        res = self.client.get("/")
+        assert res.status_code == 200
+
+    def test_invalid_path(self):
+        res = self.client.get("/invalid")
+        assert res.status_code == 404
